refactor(renderer): split SettingsFormControls below max-lines limit
